+++ server/viame_server/viame_detection.py
@@ -11,16 +11,15 @@ from girder.models.folder import Folder
 from girder.models.item import Item
 from girder.utility import ziputil
 
-from viame_server.serializers import models, viame
-from viame_server.utils import (
+from viame_server.constants import (
     ImageMimeTypes,
     ImageSequenceType,
     VideoMimeTypes,
     VideoType,
-    getTrackData,
     safeImageRegex,
-    saveTracks,
 )
+from viame_server.serializers import models, viame
+from viame_server.utils import getTrackData, saveTracks
 
 
 class ViameDetection(Resource):
